[chore] Move LoggerWithout temporarily to internal package
