+++ cmd/otelcorecol/builder-config.yaml
@@ -87,6 +87,7 @@ replaces:
   - go.opentelemetry.io/collector/featuregate => ../../featuregate
   - go.opentelemetry.io/collector/internal/memorylimiter => ../../internal/memorylimiter
   - go.opentelemetry.io/collector/internal/fanoutconsumer => ../../internal/fanoutconsumer
+  - go.opentelemetry.io/collector/internal/telemetry => ../../internal/telemetry
   - go.opentelemetry.io/collector/internal/sharedcomponent => ../../internal/sharedcomponent
   - go.opentelemetry.io/collector/otelcol => ../../otelcol
   - go.opentelemetry.io/collector/pdata => ../../pdata
