+++ versions.yaml
@@ -27,6 +27,7 @@ module-sets:
       - go.opentelemetry.io/collector/internal/memorylimiter
       - go.opentelemetry.io/collector/internal/fanoutconsumer
       - go.opentelemetry.io/collector/internal/sharedcomponent
+      - go.opentelemetry.io/collector/internal/telemetry
       - go.opentelemetry.io/collector/cmd/builder
       - go.opentelemetry.io/collector/cmd/mdatagen
       - go.opentelemetry.io/collector/component
